feat: add wait_for_text keyword (#89)
feat: add match_text keyword

context on why I implemented this:
After doing an lvm or zfs encrypted installation, there is a graphical
screen prompting for the users passphrase to decrypt the disk.
This screen can vary in its resolution from machine to machine, making
matching screenshots bothersome, requiring variants.

So, I implemented match_text, so I can simply wait for this phrase:
"Please unlock"
to be seen on the HDMI.

This commit comes complete with unit tests for the new keyword also.