[New] VNC platform (#76)
feat: Add Vnc backend

This commit adds a new backend to yarf: Vnc

The Vnc backend can be used with any kind of VM that provides a VNC
server. I've predominantly used it with qemu.

The following methods are implemented:
- Mouse clicking
- Mouse pressing
- Keyboard input
- Grabbing screenshots

The commit comes complete with the necessary unit tests, as well as
a bit of documentation to help users use this backend.