Suite parser symlinks (#109)
* suite-parser: follow symlinks when parsing assets

This commit amends the logic in yarf/rf_libraries/suite_parser.py to
also follow symlinks in the test suite directory. Unfortunately, due to
the pinned python version in yarf, there's no nicer way to do this than
importing os and using os.walk. In the current python version (3.10.15),
the pathlib library does not offer the necessary functionality.

In later python versions, the usage of os.walk can be replaced with:
- Path("/my/path").rglob("*", recurse_symlinks=True)
- Path("/my/path").walk(follow_symlinks=True)

* README.md: Install deb dependencies also