fix: `yarf` command line should exit with suite.run exit code (#135)
I noticed when running the jenkins job which uses yarf with the Vnc
backend, that even when the test failed, yarf would exit 0. The wrapper
script being used around yarf checks the exit code of yarf to decipher
whether a test has passed or failed, and was giving false positives,
which is very much less than ideal.

This commit modifies main.py and the tests for main.py to exit with
whatever code is returned from suite.run.