snap: Use gnome extension as base
Yarf requires already quite a lot of common desktop libraries that are
part of the gnome extension, so use it instead of duplicate such libs
and data files everywhere.

This will also help for future plans of using more modern UI for the
interactive selector.

As per this the size of the snap is reduced substantially