Fix the shifted keys issue in VNC with QEMU
Replace key_codes dictionary (mapping a keyboard key to a key-code) with
a key_sequences dictionary (mapping a keyboard key to a sequence of
key-codes), and re-work the _write method to hold and release the
sequence of key-codes as required.

Reading the spec it *appears* that QEMU's VNC is at fault here, but I
suspect they had little choice but to implement things as passing
through key-codes verbatim rather than relying on the character in the
message as the VNC spec requires (given QEMU has no idea what OS is
running within it and therefore what locale/keymap etc there is).

This commit also introduces a couple of extra key aliases that seemed
fairly obvious to me (Ctrl for Control, Enter for Return).