[New] Include screenshots from matching text in video
We were only including the screenshots from the match_all and match_any
keywords in the video. The screenshots taken when matching text, i.e.
match_text, find_text, and read_text were not saved and used in the
video. In effect, tests which only use OCR instead of image matching did
not log a video at all.